WIP: feat: support for per-tensor quant types
Allows changing weight quants by regex, with a similar syntax to
the llama.cpp tensor overrides. This is useful for instance to reduce
unet memory requirements without changing the more sensitive clip
weights.

I only did the conversion part for now, but I intend to adapt the inference
code afterwards.

BTW, if we consider a plain quant name as matching any tensor, this could
reuse the --type command line argument instead of demanding another one.
Same thing for the type parameters for conversion and inference, which
in principle could be simply changed to a const char*.